+++ app/src/main/java/com/example/swapit/domain/model/shopping/detail/ShoppingDetailData.kt
@@ -1,16 +1,16 @@
 package com.example.swapit.domain.model.shopping.detail
 
 data class ShoppingDetailData(
-    val goodsId: Int,
+    val goodsId: Long,
     val user: ShoppingDetailUser,
     val category: String,
     val title: String,
-    val price: Int,
+    val price: Long,
     val quality: String,
     val content: String,
     val goodsTradeStatus: String,
     val placeName: String,
-    val viewCount: Int,
+    val viewCount: Long,
     val imageUri: List<ShoppingDetailImage>,
     val createdAt: String,
 )
